requirements.txt
Flask==2.0.3
flask-apscheduler==1.12.3
python-dotenv==0.19.2
torch==1.10.1
transformers==4.12.5
scikit-learn==1.0.2